Replaced lvgl with lv_bindings submodule. lv_bindings contains the Micropython bindings and lvgl submodule (Micropython bindings was split from lvgl).
